+++ src/hmc/train/global_classifier/train_global_baseline.py
